projects/3-stock-assistant/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
yfinance>=0.2.0
